==> ivca/01_ppl_cleaning.ipynb <==
import json
import pandas as pd


df = pd.read_json('01_ppl_raw_upraveny.json')
df.head(5)


df.columns

df[['latitude', 'longitude']] = df['gps'].apply(pd.Series)

df.head(5)

df.drop(columns='gps', inplace=True)

df.head(5)

df.info()

df['accessPointType'].value_counts()

alzaboxy_ppl = df[df['accessPointType'] == "AlzaBox"]
alzaboxy_ppl

alzaboxy_ppl.to_csv('data_alzabox_ppl.csv', index=False, encoding='utf-8')

ppl_box = df[df['accessPointType'] == "ParcelBox"]
ppl_box

ppl_box.to_csv('data_ppl_box.csv', index=False, encoding='utf-8')

ppl_shop = df[df['accessPointType'] == "ParcelShop"]
ppl_shop

ppl_shop.to_csv('data_ppl_shop.csv', index=False, encoding='utf-8')
