==> ivca/01_ppl_cleaning_box.ipynb <==
import pandas as pd

df = pd.read_csv('data_ppl_box.csv')

df.head(5)

df[['box_type', 'city_abbr', 'street', 'note']] = df['name'].str.extract(r'(\w+)\s+(\w+)\s+(.*?)(?:\s+\((.*?)\))?$')

df.head(5)

df.to_csv('data_ppl_box_clean.csv', index=False, encoding='utf-8')

df['city_abbr'].unique()

df['city_abbr'].value_counts()
