==> ivca/02_ppl_test.ipynb <==
import pandas as pd

df = pd.read_csv('data_ppl_box_clean.csv')

df.head(5)

df["slouceny"] = '"' + df["latitude"].astype(str) + ',' + df["longitude"].astype(str) + '",'

df.head(5)

df.to_csv('test_ppl_vse.csv', index=False, encoding='utf-8')
