==> ivca/03_ppl_duplikaty.ipynb <==
import pandas as pd

df = pd.read_csv('test_ppl_vse.csv')

df.head(5)

radky_s_nan = df.isna().any(axis=1).sum()
print(f"{radky_s_nan} řádků obsahuje alespoň jednu NaN hodnotu.")

df_nan_radky = df[df.isna().any(axis=1)]
print(df_nan_radky)

df_clean = df.dropna()

puvodni_pocet = len(df)
novy_pocet = len(df_clean)
print(f"Odstraněno {puvodni_pocet - novy_pocet} řádků s nulovými hodnotami.")

df_clean.info

df_unique = df_clean.drop_duplicates(subset=["id", "name"])

novy_pocet = len(df_unique)
print(novy_pocet)

df_unique.to_csv('unikatni_ppl_vse.csv', index=False, encoding='utf-8')
