==> ivca/04_ppl_cleaning_new_adresses.ipynb <==
import pandas as pd

df = pd.read_csv('unikatni_ppl_vse.csv')

df.head(5)

new_df = df['address'].str.split(',', expand=True)

new_df.head(5)

new_df.to_csv('ppl_adresy.csv', index=False, encoding='utf-8')

df.to_csv('unikatni_ppl_pokus2.csv', index=False, encoding='utf-8')
